crisissync-backend/requirements.txt
# CrisisSync Backend — Python Dependencies
# Install with: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# ── Auth & Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# ── AI ────────────────────────────────────────────────────────────────────────
anthropic==0.40.0

# ── Cache ─────────────────────────────────────────────────────────────────────
redis==5.2.1
hiredis==3.1.0

# ── Background Tasks ──────────────────────────────────────────────────────────
celery==5.4.0
celery[redis]==5.4.0

# ── Notifications ─────────────────────────────────────────────────────────────
sendgrid==6.11.0

# ── Utils ─────────────────────────────────────────────────────────────────────
python-dateutil==2.9.0
httpx==0.27.2
Pillow==11.0.0

# ── AWS / File Upload ─────────────────────────────────────────────────────────
boto3==1.35.80

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
factory-boy==3.3.1
aiosqlite==0.20.0
